Add files via upload
My first commit, This is my individual semester project.
All the above codes are original and written by myself. The idea is to create a site that allow individuals to 
(1) construct a public or semi-public profile within a bounded system, 
(2) articulate a list of other users with whom they share a connection, and 
(3) view and traverse their list of connections and those made by others within the system.